separated period parse/serialize code do be able to use outside context of configtree.
