+++ pyhocon/config_parser.py
@@ -1,21 +1,20 @@
 import codecs
 import contextlib
 import copy
-import itertools
 import logging
 import os
 import re
 import socket
 import sys
-from datetime import timedelta
 
 import pyparsing
-
 from pyparsing import (Forward, Group, Keyword, Literal, Optional,
                        ParserElement, ParseSyntaxException, QuotedString,
                        Regex, SkipTo, StringEnd, Suppress, TokenConverter,
                        Word, ZeroOrMore, alphanums, alphas8bit, col, lineno,
-                       replaceWith, Or, nums, White, WordEnd)
+                       replaceWith)
+
+from pyhocon.period_parser import get_period_expr
 
 # Fix deepcopy issue with pyparsing
 if sys.version_info >= (3, 8):
@@ -27,6 +26,7 @@ if sys.version_info >= (3, 8):
         except KeyError:
             return ""
 
+
     pyparsing.ParseResults.__getattr__ = fixed_get_attr
 
 from pyhocon.config_tree import (ConfigInclude, ConfigList, ConfigQuotedString,
@@ -34,7 +34,6 @@ from pyhocon.config_tree import (ConfigInclude, ConfigList, ConfigQuotedString,
                                  ConfigUnquotedString, ConfigValues, NoneValue)
 from pyhocon.exceptions import (ConfigException, ConfigMissingException,
                                 ConfigSubstitutionException)
-
 
 use_urllib2 = False
 try:
@@ -62,11 +61,11 @@ if sys.version_info < (3, 5):
 else:
     from glob import glob
 
-
 # Fix deprecated warning with 'imp' library and Python 3.4+.
 # See: https://github.com/chimpler/pyhocon/issues/248
 if sys.version_info >= (3, 4):
     import importlib.util
+
 
     def find_package_dirs(name):
         spec = importlib.util.find_spec(name)
@@ -80,11 +79,12 @@ else:
     import imp
     import importlib
 
+
     def find_package_dirs(name):
         return [imp.find_module(name)[1]]
 
-
 logger = logging.getLogger(__name__)
+
 
 #
 # Substitution Defaults
@@ -105,24 +105,6 @@ class NO_SUBSTITUTION(object):
 
 class STR_SUBSTITUTION(object):
     pass
-
-
-def period(period_value, period_unit):
-    try:
-        from dateutil.relativedelta import relativedelta as period_impl
-    except Exception:
-        from datetime import timedelta as period_impl
-
-    if period_unit == 'nanoseconds':
-        period_unit = 'microseconds'
-        period_value = int(period_value / 1000)
-
-    arguments = dict(zip((period_unit,), (period_value,)))
-
-    if period_unit == 'milliseconds':
-        return timedelta(**arguments)
-
-    return period_impl(**arguments)
 
 
 class ConfigFactory(object):
@@ -241,42 +223,6 @@ class ConfigParser(object):
         '\\"': '"',
     }
 
-    period_type_map = {
-        'nanoseconds': ['ns', 'nano', 'nanos', 'nanosecond', 'nanoseconds'],
-
-        'microseconds': ['us', 'micro', 'micros', 'microsecond', 'microseconds'],
-        'milliseconds': ['ms', 'milli', 'millis', 'millisecond', 'milliseconds'],
-        'seconds': ['s', 'second', 'seconds'],
-        'minutes': ['m', 'minute', 'minutes'],
-        'hours': ['h', 'hour', 'hours'],
-        'weeks': ['w', 'week', 'weeks'],
-        'days': ['d', 'day', 'days'],
-
-    }
-
-    optional_period_type_map = {
-        'months': ['mo', 'month', 'months'],  # 'm' from hocon spec removed. conflicts with minutes syntax.
-        'years': ['y', 'year', 'years']
-    }
-
-    supported_period_map = None
-
-    @classmethod
-    def get_supported_period_type_map(cls):
-        if cls.supported_period_map is None:
-            cls.supported_period_map = {}
-            cls.supported_period_map.update(cls.period_type_map)
-
-            try:
-                from dateutil import relativedelta
-
-                if relativedelta is not None:
-                    cls.supported_period_map.update(cls.optional_period_type_map)
-            except Exception:
-                pass
-
-        return cls.supported_period_map
-
     @classmethod
     def parse(cls, content, basedir=None, resolve=True, unresolved_value=DEFAULT_SUBSTITUTION):
         """parse a HOCON content
@@ -314,16 +260,6 @@ class ConfigParser(object):
                 return int(n, 10)
             except ValueError:
                 return float(n)
-
-        def convert_period(tokens):
-            period_value = int(tokens.value)
-            period_identifier = tokens.unit
-
-            period_unit = next((single_unit for single_unit, values
-                                in cls.get_supported_period_type_map().items()
-                                if period_identifier in values))
-
-            return period(period_value, period_unit)
 
         # ${path} or ${?path} for optional substitution
         SUBSTITUTION_PATTERN = r"\$\{(?P<optional>\?)?(?P<variable>[^}]+)\}(?P<ws>[ \t]*)"
@@ -452,18 +388,6 @@ class ConfigParser(object):
             number_expr = Regex(r'[+-]?(\d*\.\d+|\d+(\.\d+)?)([eE][+\-]?\d+)?(?=$|[ \t]*([\$\}\],#\n\r]|//))',
                                 re.DOTALL).setParseAction(convert_number)
 
-            # Flatten the list of lists with unit strings.
-            period_types = list(itertools.chain(*cls.get_supported_period_type_map().values()))
-            # `Or()` tries to match the longest expression if more expressions
-            # are matching. We employ this to match e.g.: 'weeks' so that we
-            # don't end up with 'w' and 'eeks'. Note that 'weeks' but also 'w'
-            # are valid unit identifiers.
-            # Allow only spaces as a valid separator between value and unit.
-            # E.g. \t as a separator is invalid: '10<TAB>weeks'.
-            period_expr = (
-                Word(nums)('value') + ZeroOrMore(White(ws=' ')).suppress() + Or(period_types)('unit') + WordEnd(alphanums).suppress()
-            ).setParseAction(convert_period)
-
             # multi line string using """
             # Using fix described in http://pyparsing.wikispaces.com/share/view/3778969
             multiline_string = Regex('""".*?"*"""', re.DOTALL | re.UNICODE).setParseAction(parse_multi_string)
@@ -478,7 +402,7 @@ class ConfigParser(object):
             substitution_expr = Regex(r'[ \t]*\$\{[^\}]+\}[ \t]*').setParseAction(create_substitution)
             string_expr = multiline_string | quoted_string | unquoted_string
 
-            value_expr = period_expr | number_expr | true_expr | false_expr | null_expr | string_expr
+            value_expr = get_period_expr() | number_expr | true_expr | false_expr | null_expr | string_expr
 
             include_content = (
                 quoted_string | ((Keyword('url') | Keyword('file') | Keyword('package')) - Literal('(').suppress() - quoted_string - Literal(')').suppress())
