+++ pyhocon/converter.py
@@ -1,7 +1,6 @@
 import json
 import re
 import sys
-from datetime import timedelta
 
 from pyhocon import ConfigFactory
 from pyhocon.config_tree import ConfigQuotedString
@@ -9,7 +8,7 @@ from pyhocon.config_tree import ConfigSubstitution
 from pyhocon.config_tree import ConfigTree
 from pyhocon.config_tree import ConfigValues
 from pyhocon.config_tree import NoneValue
-
+from pyhocon.period_serializer import timedelta_to_str, is_timedelta_like, timedelta_to_hocon
 
 try:
     basestring
@@ -61,8 +60,8 @@ class HOCONConverter(object):
                     )
                 lines += ',\n'.join(bet_lines)
                 lines += '\n{indent}]'.format(indent=''.rjust(level * indent, ' '))
-        elif cls._is_timedelta_like(config):
-            lines += cls._timedelta_to_str(config)
+        elif is_timedelta_like(config):
+            lines += timedelta_to_str(config)
         elif isinstance(config, basestring):
             lines = json.dumps(config, ensure_ascii=False)
         elif config is None or isinstance(config, NoneValue):
@@ -138,8 +137,8 @@ class HOCONConverter(object):
                 lines = '"""{value}"""'.format(value=config.value)  # multilines
             else:
                 lines = '"{value}"'.format(value=cls.__escape_string(config.value))
-        elif cls._is_timedelta_like(config):
-            lines += cls._timedelta_to_hocon(config)
+        elif is_timedelta_like(config):
+            lines += timedelta_to_hocon(config)
         elif config is None or isinstance(config, NoneValue):
             lines = 'null'
         elif config is True:
@@ -181,8 +180,8 @@ class HOCONConverter(object):
                     bet_lines.append('{indent}- {value}'.format(indent=''.rjust(level * indent, ' '),
                                                                 value=cls.to_yaml(item, compact, indent, level + 1)))
                 lines += '\n'.join(bet_lines)
-        elif cls._is_timedelta_like(config):
-            lines += cls._timedelta_to_str(config)
+        elif is_timedelta_like(config):
+            lines += timedelta_to_str(config)
         elif isinstance(config, basestring):
             # if it contains a \n then it's multiline
             lines = config.split('\n')
@@ -223,8 +222,8 @@ class HOCONConverter(object):
             for index, item in enumerate(config):
                 if item is not None:
                     lines.append(cls.to_properties(item, compact, indent, stripped_key_stack + [str(index)]))
-        elif cls._is_timedelta_like(config):
-            lines.append('.'.join(stripped_key_stack) + ' = ' + cls._timedelta_to_str(config))
+        elif is_timedelta_like(config):
+            lines.append('.'.join(stripped_key_stack) + ' = ' + timedelta_to_str(config))
         elif isinstance(config, basestring):
             lines.append('.'.join(stripped_key_stack) + ' = ' + escape_value(config))
         elif config is True:
@@ -292,45 +291,3 @@ class HOCONConverter(object):
     def __escape_string(cls, string):
         return re.sub(r'[\x00-\x1F"\\]', cls.__escape_match, string)
 
-    @classmethod
-    def _is_timedelta_like(cls, config):
-        return isinstance(config, timedelta) or relativedelta is not None and isinstance(config, relativedelta)
-
-    @classmethod
-    def _timedelta_to_str(cls, config):
-        if isinstance(config, relativedelta):
-            time_delta = cls._relative_delta_to_timedelta(config)
-        else:
-            time_delta = config
-        return str(int(time_delta.total_seconds() * 1000))
-
-    @classmethod
-    def _timedelta_to_hocon(cls, config):
-        """
-        :type config: timedelta
-        """
-        if relativedelta is not None and isinstance(config, relativedelta):
-            if config.hours > 0:
-                return str(config.hours) + ' hours'
-            elif config.minutes > 0:
-                return str(config.minutes) + ' minutes'
-
-        if config.days > 0:
-            return str(config.days) + ' days'
-        elif config.seconds > 0:
-            return str(config.seconds) + ' seconds'
-        elif config.microseconds > 0:
-            return str(config.microseconds) + ' microseconds'
-        else:
-            return '0 seconds'
-
-    @classmethod
-    def _relative_delta_to_timedelta(cls, relative_delta):
-        """
-        :type relative_delta: relativedelta
-        """
-        return timedelta(days=relative_delta.days,
-                         hours=relative_delta.hours,
-                         minutes=relative_delta.minutes,
-                         seconds=relative_delta.seconds,
-                         microseconds=relative_delta.microseconds)
